# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SearchResult {
  id: string;
  voterName: string;
  relativeName: string;
  relativeLabel?: string;
  /** @nullable */
  epicNumber?: string | null;
  /** @nullable */
  serialNumber: string | null;
  /** @nullable */
  houseNumber: string | null;
  /** @nullable */
  age: number | null;
  /** @nullable */
  gender: string | null;
  partNumber: string;
  pageNumber: number;
  pdfName: string;
  pdfId: string;
  confidence: number;
  score: number;
  matchedBy?: string[];
}

export type SearchResponseFilters = {
  /** @nullable */
  name: string | null;
  /** @nullable */
  relativeName: string | null;
  /** @nullable */
  epicNumber: string | null;
};

export interface SearchResponse {
  results: SearchResult[];
  total: number;
  page: number;
  pageSize: number;
  query: string;
  filters: SearchResponseFilters;
  indexStatus: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  indexProgress: number;
}

export interface SearchSuggestion {
  label: string;
  value: string;
  kind: string;
}

export interface PdfAsset {
  id: string;
  name: string;
  sizeBytes: number;
  pageCount: number;
  uploadedAt: string;
  status: string;
  indexedRecords: number;
  fileUrl?: string;
}

export interface PdfRenameInput {
  /** @minLength 1 */
  name: string;
}

export interface IndexStats {
  totalPdfs: number;
  totalRecords: number;
  status: string;
  /** @nullable */
  lastIndexedAt: string | null;
  indexVersion: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
  indexedPdfs: number;
  failedPdfs: number;
  ocrPages: number;
  recordsWithEpic: number;
  recordsWithRelativeName: number;
  recordsWithDetails: number;
  warnings: string[];
}

export type SearchElectoralRollParams = {
/**
 * @minLength 1
 */
q: string;
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 50
 */
pageSize?: number;
};

export type GetSearchSuggestionsParams = {
/**
 * @minLength 2
 */
q: string;
};

export type ListPdfAssetsParams = {
q?: string;
};

export type UploadPdfParams = {
filename: string;
};