# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: search
    description: Public electoral roll search
  - name: admin
    description: PDF and search index management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /search:
    get:
      operationId: searchElectoralRoll
      tags: [search]
      summary: Search indexed electoral roll records
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "400":
          description: Invalid search query
  /suggestions:
    get:
      operationId: getSearchSuggestions
      tags: [search]
      summary: Get search suggestions from indexed names and EPIC numbers
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
      responses:
        "200":
          description: Suggestions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SearchSuggestion"
  /admin/pdfs:
    get:
      operationId: listPdfAssets
      tags: [admin]
      summary: List managed electoral roll PDFs
      parameters:
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Managed PDFs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PdfAsset"
    post:
      operationId: uploadPdf
      tags: [admin]
      summary: Upload one PDF and start indexing
      parameters:
        - name: filename
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "201":
          description: PDF uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PdfAsset"
  /admin/pdfs/{id}:
    patch:
      operationId: renamePdf
      tags: [admin]
      summary: Rename a managed PDF
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PdfRenameInput"
      responses:
        "200":
          description: Renamed PDF
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PdfAsset"
    delete:
      operationId: deletePdf
      tags: [admin]
      summary: Delete a managed PDF and its index entries
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: PDF deleted
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get PDF and index health
      responses:
        "200":
          description: Index statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IndexStats"
  /admin/index/rebuild:
    post:
      operationId: rebuildSearchIndex
      tags: [admin]
      summary: Rebuild the local search index in the background
      responses:
        "202":
          description: Rebuild started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IndexStats"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    SearchResult:
      type: object
      required: [id, voterName, relativeName, partNumber, pageNumber, pdfName, pdfId, score, serialNumber, houseNumber, age, gender, confidence]
      properties:
        id: { type: string }
        voterName: { type: string }
        relativeName: { type: string }
        relativeLabel: { type: string }
        epicNumber: { type: ["string", "null"] }
        serialNumber: { type: ["string", "null"] }
        houseNumber: { type: ["string", "null"] }
        age: { type: ["number", "null"] }
        gender: { type: ["string", "null"] }
        partNumber: { type: string }
        pageNumber: { type: number }
        pdfName: { type: string }
        pdfId: { type: string }
        confidence: { type: number }
        score: { type: number }
        matchedBy: { type: array, items: { type: string } }
    SearchResponse:
      type: object
      required: [results, total, page, pageSize, query, filters, indexStatus, indexProgress]
      properties:
        results: { type: array, items: { $ref: "#/components/schemas/SearchResult" } }
        total: { type: number }
        page: { type: number }
        pageSize: { type: number }
        query: { type: string }
        filters:
          type: object
          required: [name, relativeName, epicNumber]
          properties:
            name: { type: ["string", "null"] }
            relativeName: { type: ["string", "null"] }
            epicNumber: { type: ["string", "null"] }
        indexStatus: { type: string }
        indexProgress: { type: number, minimum: 0, maximum: 100 }
    SearchSuggestion:
      type: object
      required: [label, value, kind]
      properties:
        label: { type: string }
        value: { type: string }
        kind: { type: string }
    PdfAsset:
      type: object
      required: [id, name, sizeBytes, pageCount, uploadedAt, status, indexedRecords]
      properties:
        id: { type: string }
        name: { type: string }
        sizeBytes: { type: number }
        pageCount: { type: number }
        uploadedAt: { type: string, format: date-time }
        status: { type: string }
        indexedRecords: { type: number }
        fileUrl: { type: string }
    PdfRenameInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
    IndexStats:
      type: object
      required: [totalPdfs, totalRecords, status, lastIndexedAt, indexVersion, progress, indexedPdfs, failedPdfs, ocrPages, recordsWithEpic, recordsWithRelativeName, recordsWithDetails, warnings]
      properties:
        totalPdfs: { type: number }
        totalRecords: { type: number }
        status: { type: string }
        lastIndexedAt: { type: ["string", "null"], format: date-time }
        indexVersion: { type: number }
        progress: { type: number, minimum: 0, maximum: 100 }
        indexedPdfs: { type: number }
        failedPdfs: { type: number }
        ocrPages: { type: number }
        recordsWithEpic: { type: number }
        recordsWithRelativeName: { type: number }
        recordsWithDetails: { type: number }
        warnings: { type: array, items: { type: string } }