# lib/api-zod/src/generated/types/indexStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface IndexStats {
  totalPdfs: number;
  totalRecords: number;
  status: string;
  /** @nullable */
  lastIndexedAt: Date | null;
  indexVersion: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
  indexedPdfs: number;
  failedPdfs: number;
  ocrPages: number;
  recordsWithEpic: number;
  recordsWithRelativeName: number;
  recordsWithDetails: number;
  warnings: string[];
}