# lib/api-zod/src/generated/types/pdfAsset.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PdfAsset {
  id: string;
  name: string;
  sizeBytes: number;
  pageCount: number;
  uploadedAt: Date;
  status: string;
  indexedRecords: number;
  fileUrl?: string;
}