# lib/api-zod/src/generated/types/searchResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SearchResponseFilters } from './searchResponseFilters';
import type { SearchResult } from './searchResult';

export interface SearchResponse {
  results: SearchResult[];
  total: number;
  page: number;
  pageSize: number;
  query: string;
  filters: SearchResponseFilters;
  indexStatus: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  indexProgress: number;
}