# lib/api-zod/src/generated/types/searchResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SearchResult {
  id: string;
  voterName: string;
  relativeName: string;
  relativeLabel?: string;
  /** @nullable */
  epicNumber?: string | null;
  /** @nullable */
  serialNumber: string | null;
  /** @nullable */
  houseNumber: string | null;
  /** @nullable */
  age: number | null;
  /** @nullable */
  gender: string | null;
  partNumber: string;
  pageNumber: number;
  pdfName: string;
  pdfId: string;
  confidence: number;
  score: number;
  matchedBy?: string[];
}